requirements.txt
# Interface Web
streamlit>=1.35.0
plotly>=5.20.0

# Manipulação de Dados e Ficheiros Excel
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.2

# Estruturação e Validação de Dados (Schemas)
pydantic>=2.4.0

# Integração com LLM (OpenRouter/OpenAI)
langchain-openai>=0.1.0
langchain-core>=0.1.0

# Graph Database e Raciocínio Baseado em Grafos
langgraph

# Gestão de Variáveis de Ambiente (.env)
python-dotenv>=1.0.0
